mobile viewport fixes but also bad news
  - ✅ Removed the "Match • Heal • Succeed" subtitle with a simple sed command
  - ✅ Fixed mobile viewport issues for iPhone 13 mini
  - ✅ Added drag/swipe functionality for better mobile experience

Unfortunately, I also realized that somewhere along the way, probably in my first interaction with Claude Code, taking from Repl.it to my GitHub, it minified everything into this JavaScript file. So all of the very simple copy changes now have to be done inside of this.js which is just awful. I looked into how to switch over to a fresh directory downloaded from Repl.it which has a very, very simplified page structure or file structure. And there were just so many things that I didn't want to redo. So we will put in this fix and then pump that for the future.